LEL-014 refactor(database): organize option and mood journal entities
