+++ core/data/src/main/java/io/github/faening/lello/core/data/repository/EmotionOptionRepository.kt
@@ -1,7 +1,7 @@
 package io.github.faening.lello.core.data.repository
 
 import io.github.faening.lello.core.database.dao.EmotionOptionDao
-import io.github.faening.lello.core.database.model.EmotionOptionEntity
+import io.github.faening.lello.core.database.model.option.EmotionOptionEntity
 import io.github.faening.lello.core.model.journal.EmotionOption
 import jakarta.inject.Inject
