+++ core/data/src/main/java/io/github/faening/lello/core/data/repository/HealthOptionRepository.kt
@@ -1,7 +1,7 @@
 package io.github.faening.lello.core.data.repository
 
 import io.github.faening.lello.core.database.dao.HealthOptionDao
-import io.github.faening.lello.core.database.model.HealthOptionEntity
+import io.github.faening.lello.core.database.model.option.HealthOptionEntity
 import io.github.faening.lello.core.model.journal.HealthOption
 import jakarta.inject.Inject
