+++ core/data/src/main/java/io/github/faening/lello/core/data/repository/SocialOptionRepository.kt
@@ -1,7 +1,7 @@
 package io.github.faening.lello.core.data.repository
 
 import io.github.faening.lello.core.database.dao.SocialOptionDao
-import io.github.faening.lello.core.database.model.SocialOptionEntity
+import io.github.faening.lello.core.database.model.option.SocialOptionEntity
 import io.github.faening.lello.core.model.journal.SocialOption
 import jakarta.inject.Inject
