+++ core/database/src/main/java/io/github/faening/lello/core/database/LelloDatabase.kt
@@ -18,26 +18,26 @@ import io.github.faening.lello.core.database.dao.SensationOptionDao
 import io.github.faening.lello.core.database.dao.SleepQualityOptionDao
 import io.github.faening.lello.core.database.dao.SocialOptionDao
 import io.github.faening.lello.core.database.dao.SleepActivityOptionDao
-import io.github.faening.lello.core.database.model.ClimateOptionEntity
-import io.github.faening.lello.core.database.model.DosageFormOptionEntity
-import io.github.faening.lello.core.database.model.AppetiteOptionEntity
-import io.github.faening.lello.core.database.model.EmotionOptionEntity
-import io.github.faening.lello.core.database.model.FoodOptionEntity
-import io.github.faening.lello.core.database.model.HealthOptionEntity
+import io.github.faening.lello.core.database.model.option.ClimateOptionEntity
+import io.github.faening.lello.core.database.model.option.DosageFormOptionEntity
+import io.github.faening.lello.core.database.model.option.AppetiteOptionEntity
+import io.github.faening.lello.core.database.model.option.EmotionOptionEntity
+import io.github.faening.lello.core.database.model.option.FoodOptionEntity
+import io.github.faening.lello.core.database.model.option.HealthOptionEntity
 import io.github.faening.lello.core.database.model.JournalCategoryEntity
-import io.github.faening.lello.core.database.model.LocationOptionEntity
-import io.github.faening.lello.core.database.model.MealOptionEntity
-import io.github.faening.lello.core.database.model.PortionOptionEntity
-import io.github.faening.lello.core.database.model.MoodJournalEntity
-import io.github.faening.lello.core.database.model.MoodJournalEntityClimateOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntityEmotionOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntityHealthOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntityLocationOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntitySocialOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.SensationOptionEntity
-import io.github.faening.lello.core.database.model.SleepQualityOptionEntity
-import io.github.faening.lello.core.database.model.SocialOptionEntity
-import io.github.faening.lello.core.database.model.SleepActivityOptionEntity
+import io.github.faening.lello.core.database.model.option.LocationOptionEntity
+import io.github.faening.lello.core.database.model.option.MealOptionEntity
+import io.github.faening.lello.core.database.model.option.PortionOptionEntity
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntity
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityClimateOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityEmotionOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityHealthOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityLocationOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntitySocialOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.option.SensationOptionEntity
+import io.github.faening.lello.core.database.model.option.SleepQualityOptionEntity
+import io.github.faening.lello.core.database.model.option.SocialOptionEntity
+import io.github.faening.lello.core.database.model.option.SleepActivityOptionEntity
 import io.github.faening.lello.core.database.util.DateConverter
 import io.github.faening.lello.core.database.util.InstantConverters
 import io.github.faening.lello.core.database.util.JournalMoodTypeConverter
