+++ core/database/src/main/java/io/github/faening/lello/core/database/dao/MoodJournalDao.kt
@@ -6,12 +6,12 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Transaction
-import io.github.faening.lello.core.database.model.MoodJournalEntity
-import io.github.faening.lello.core.database.model.MoodJournalEntityClimateOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntityEmotionOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntityHealthOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntityLocationOptionEntityCrossRef
-import io.github.faening.lello.core.database.model.MoodJournalEntitySocialOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntity
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityClimateOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityEmotionOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityHealthOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntityLocationOptionEntityCrossRef
+import io.github.faening.lello.core.database.model.moodjournal.MoodJournalEntitySocialOptionEntityCrossRef
 import io.github.faening.lello.core.domain.repository.MoodJournalResources
 
 @Dao
