+++ core/database/src/main/java/io/github/faening/lello/core/database/dao/PortionOptionDao.kt
@@ -7,7 +7,7 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Update
-import io.github.faening.lello.core.database.model.PortionOptionEntity
+import io.github.faening.lello.core.database.model.option.PortionOptionEntity
 import io.github.faening.lello.core.domain.repository.OptionResources
 import kotlinx.coroutines.flow.Flow
