+++ core/database/src/main/java/io/github/faening/lello/core/database/dao/SocialOptionDao.kt
@@ -7,7 +7,7 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Update
-import io.github.faening.lello.core.database.model.SocialOptionEntity
+import io.github.faening.lello.core.database.model.option.SocialOptionEntity
 import io.github.faening.lello.core.domain.repository.OptionResources
 import kotlinx.coroutines.flow.Flow
