+++ core/database/src/main/java/io/github/faening/lello/core/database/model/moodjournal/MoodJournalEntityWithOptions.kt
@@ -1,8 +1,13 @@
-package io.github.faening.lello.core.database.model
+package io.github.faening.lello.core.database.model.moodjournal
 
 import androidx.room.Embedded
 import androidx.room.Junction
 import androidx.room.Relation
+import io.github.faening.lello.core.database.model.option.ClimateOptionEntity
+import io.github.faening.lello.core.database.model.option.EmotionOptionEntity
+import io.github.faening.lello.core.database.model.option.HealthOptionEntity
+import io.github.faening.lello.core.database.model.option.LocationOptionEntity
+import io.github.faening.lello.core.database.model.option.SocialOptionEntity
 
 data class MoodJournalEntityWithOptions(
     @Embedded val entry: MoodJournalEntity,
