+++ core/database/src/main/java/io/github/faening/lello/core/database/model/option/SensationOptionEntity.kt
@@ -1,4 +1,4 @@
-package io.github.faening.lello.core.database.model
+package io.github.faening.lello.core.database.model.option
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
