+++ core/database/src/main/java/io/github/faening/lello/core/database/seed/SocialOptionSeed.kt
@@ -1,6 +1,6 @@
 package io.github.faening.lello.core.database.seed
 
-import io.github.faening.lello.core.database.model.SocialOptionEntity
+import io.github.faening.lello.core.database.model.option.SocialOptionEntity
 
 object SocialOptionSeed : Seed<SocialOptionEntity> {
     override val data = listOf(
